Integer indexing redesign: one descriptor class decided in basis.m, sparse single basis storage, class-inheriting lin2lm and lm2lin
- basis.m picks min_int_type(max(mults)^2-1,'signed') once; the state lists, the subgraph blocks, and the projections used by the filters carry that class; blocks are embedded as sparse single and the stacked descriptor is deduplicated and sorted by one unique call; the byte rule, typecast, dense copies, and double() conversions are gone
- lin2lm: the square root is the only floating-point step, projections are L.^2-I+L in the class of the input, and a posterior check catches a root rounded across an integer; lm2lin is L.^2-M+L with a grumbler bound on the rank for integer classes
- superop.m matches the passive columns on the sparse single descriptor without a dense copy
- relaxation.m and homospoil.m promote projections to double where they multiply Larmor frequencies, because single-precision sums of +-omega do not cancel exactly; symmetry.m hands spsortrows the sparse double its MEX requires; min_int_type accepts a maximum of zero for ghost-only systems
- knowledge entries, skill reference, and the indexing test suite updated

+++ kernel/homospoil.m
@@ -76,7 +76,7 @@ switch zqc_flag
     case 'keep'
         
         % Find the states that have zero carrier frequency and kill everything else
-        rho(abs(sum(repmat(spin_system.inter.basefrqs,size(spin_system.bas.basis,1),1).*M,2))>1e-6,:)=0;
+        rho(abs(sum(repmat(spin_system.inter.basefrqs,size(spin_system.bas.basis,1),1).*double(M),2))>1e-6,:)=0;
     
     case 'destroy'
